verbatim-parser /requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
algoliasearch==2.6.0
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
Automat==20.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
constantly==15.1.0
cryptography==36.0.0
cssselect==1.1.0
dill==0.3.4
fire==0.4.0
Flask==2.0.2
Flask-Cors==3.0.10
frozenlist==1.2.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.8.2
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
lxml==4.6.4
MarkupSafe==2.0.1
multidict==5.2.0
multiprocess==0.70.12.2
parse==1.19.0
parsel==1.6.0
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==21.0.0
pyquery==1.4.3
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.19.2
queuelib==1.6.2
requests==2.26.0
Scrapy==2.5.1
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.1
termcolor==1.1.0
Twisted==21.7.0
typing_extensions==4.0.0
urllib3==1.26.7
w3lib==1.22.0
waitress==2.1.2
Werkzeug==2.0.2
yarl==1.7.2
zipp==3.6.0
zope.interface==5.4.0
